cr3 debug

+++ SBUnix/include/sys/physical.h
@@ -18,5 +18,5 @@ typedef struct page page_sp;
 int init_phy_page(uint32_t,uint32_t,uint32_t);
 uint32_t find_first_free();
 uint32_t find_free_pages(uint32_t);
-uint32_t allocate_page();
-uint32_t allocate_pages(uint32_t);
+uint64_t allocate_page();
+uint64_t allocate_pages(uint32_t);
